Move FirstDepleted/Ending to limiting_contact module Split enable_node_management in node_tx/node_rx, rename enable_..
