+++ src/node_manager/none.rs
@@ -6,28 +6,29 @@ use crate::{
 
 use super::NodeManager;
 
-/// Use this manager if no node management shall be considered (with or without the "enable_node_management" compilation feature).
+/// Use this manager if no node management shall be considered (with or without the node_rx, node_tx, and node_proc compilation feature).
 #[cfg_attr(feature = "debug", derive(Debug))]
 pub struct NoManagement {}
 
 /// This manager has no effect.
 impl NodeManager for NoManagement {
+    #[cfg(feature = "node_proc")]
     fn dry_run_process(&self, at_time: Date, _bundle: &mut Bundle) -> Date {
         return at_time;
     }
-
+    #[cfg(feature = "node_tx")]
     fn dry_run_tx(&self, _waiting_since: Date, _start: Date, _end: Date, _bundle: &Bundle) -> bool {
         true
     }
-
+    #[cfg(feature = "node_rx")]
     fn dry_run_rx(&self, _start: Date, _end: Date, _bundle: &Bundle) -> bool {
         true
     }
-
+    #[cfg(feature = "node_proc")]
     fn schedule_process(&self, at_time: Date, _bundle: &mut Bundle) -> Date {
         return at_time;
     }
-
+    #[cfg(feature = "node_tx")]
     fn schedule_tx(
         &mut self,
         _waiting_since: Date,
@@ -37,7 +38,7 @@ impl NodeManager for NoManagement {
     ) -> bool {
         true
     }
-
+    #[cfg(feature = "node_rx")]
     fn schedule_rx(&mut self, _start: Date, _end: Date, _bundle: &Bundle) -> bool {
         true
     }
