+++ src/pathfinding/limiting_contact/first_depleted.rs
