+++ src/pathfinding/limiting_contact/first_ending.rs
